Life Management
Upon starting up, the mod creates teams with different colors, and a scoreboard to keep track of the lives.

Added the `/lastlife` command, which will be the parent command for most of the last life commands (for ex. `/session` and `/givelife` will be separate)

You can now assign a random number of lives to all players / one specific player with the `/lastlife lives assignRandomLives [<target>]` command.

When a player dies, one of their lives gets subtracted, until a minimum of zero.

When a player is assigned lives, or dies, they are automatically put into their corresponding color team.