v.1.0.3
Implemented the semi-random lives choosing at the start for more diversity.

If a player is not present while the boogeyman is being chosen, then after they log back on, they have a 1/playerCount chance to become the boogeyman.

The server now broadcasts a message if the boogeyman fails to kill a player in the session.