v.1.0.9
[DOUBLE LIFE]

Implemented health and lives sharing for soulmates.

Added a custom death message for when your soulmate dies (cuz you then you also die). - Updated the resourcepack to have the message.

Added the Totem of Undying to the blacklist cuz it would not work with the health sharing. Gapples and Notch apples might follow suit, i am yet to decide on that.