v.1.1.4
[GENERAL]

Added '/lifeseries setSeries <series>' which changes the current series. from in-game. You get prompted to do this command when you first turn on the server with this mod.
This means that you no longer have to manually specify the series in the properties before starting the server - you can now change it from in game.

Added '/lives resetAll', which resets the lives of every person on the server.

[LIMITED LIFE]
When using Limited Life, the command for changing lives is changed to take in a time parameter (same as in '/session') instead of an int '/lives set/remove/add <time>'.