v.1.1.5
The datapacks are now bundled with the mod jar itself, so you don't have to copy them to the world manually.

A separate github repo was made for the Life Series resourcepack, and is automatically applied to players when they join. You do not have to specify the resourcepack in sever.properties yourself.

With these changes basically all you need to do is drop the jar file into the mods folder and thats it. Everything is done for you :)