v.1.1.9
Finished Limited Life

Added the '/claimkill' command, which lets you claim a kill for when a player dies because of you (for example from a trap). An admin can confirm this claim, and only then does the kill get properly processed.
So for example if you trap-kill someone as the Boogeyman, you can do '/claimkill Mat0u5', and if the admin confirms this kill, you will be automatically cured. Same goes for if you get rewards (like an extra life) for killing in a series.

Updated README

Bug Fixes