v.1.1.9.5
[GENERAL]
Dead players are now allowed to use '/boogeyman list'
When a player dies, a lightning bolt (harmless - no damage and no fire) is spawned where they die, instead of just playing the lightning sound.

[SECRET LIFE]
Added the '/health' command, which can be used to control how many hearts the player has. Subcommands and usage is basically the same as with '/lives'.