dev-1.2.2.59
Added the Astral Projection superpower.
Allows you to leave your body and go into spectator mode.
The "body" is a fake player, similar to the ones in the carpet mod.
If the fake player takes damage, you will be taken out of your projection.
BTW the fake players only work in 1.21.1 for SOME REASON IDK WHY